Restructured what scripts do what
Changed the SplineCurve class from an infromation container to an abstract class used as an interface. CatmullRomSpline now inherits from SplineCurve. They are interfaced by calling the function GetPosition(t) in SplineCurve. An object moving along a SplineCurve uses the component SplineCurveFollower, with a reference to a SplineCurve. These changes, in addition to making the system more logical, make it possible to translate a spline in its entirety, since it is now its own gameobject with its own transform.